Consolidate CensusResult into model project
