+++ geocoder/src/main/scala/grasshopper/geocoder/api/GeocodeFlow.scala
@@ -3,11 +3,12 @@ package grasshopper.geocoder.api
 import akka.stream.scaladsl._
 import grasshopper.client.addresspoints.AddressPointsClient
 import grasshopper.client.census.CensusClient
-import grasshopper.client.census.model.{ CensusResult, ParsedInputAddress }
+import grasshopper.client.census.model.ParsedInputAddress
 import grasshopper.client.parser.AddressParserClient
 import grasshopper.client.parser.model.ParsedAddress
 import grasshopper.geocoder.model.{ BatchGeocodeResult, AddressPointsGeocodeBatchResult, CensusGeocodeBatchResult, ParsedOutputBatchAddress }
 import grasshopper.model.addresspoints.AddressPointsResult
+import grasshopper.model.census.CensusResult
 
 import scala.concurrent.ExecutionContext
