+++ geocoder/src/main/scala/grasshopper/geocoder/model/GeocodeResult.scala
@@ -1,7 +1,7 @@
 package grasshopper.geocoder.model
 
-import grasshopper.client.census.model.CensusResult
 import grasshopper.client.parser.model.ParsedAddress
 import grasshopper.model.addresspoints.AddressPointsResult
+import grasshopper.model.census.CensusResult
 
 case class GeocodeResult(status: String, input: String, query: ParsedAddress, addressPointsService: AddressPointsResult, censusService: CensusResult)
